Tweak plugin loader code to avoid possible PHP fatal errors under certain conditions.
Closes #330.
